Changes as highlighted during meeting on 2024/10/15
We agreed to introduce an NcRestoreMode type with 2 options called:
- Rebuild
- Modify

These are used used in ValidateSetPropertiesByPath and SetPropertiesByPath to describe which restore mode the client wishes to perform as part of the request.

For a "Rebuild" restore even readonly properties can be used as information for constructing/reconstructing objects. We felt in this mode it's still important to capture if blocks are dynamic with respect to their members (if they allow their members to be added/removed) and to capture if objects can be reconstructed. For this we want to expose a new boolean field called "isRebuildable" in NcObjectPropertiesHolder.

The property traits concept is being removed:
- the NcPropertyTrait type
- NcPropertyValueHolder: traits
- ValidateSetPropertiesByPath(... propertyTraits)
- SetPropertiesByPath(... propertyTraits)

The NcRestoreValidationStatus: Excluded option needs to be removed as it was only pertinent to filtering performed as part of the property traits interaction.